restructure en.js and zh.js
