+++ config/locales/en/all.js
@@ -1,0 +1,71 @@
+export const all = {
+    defaultText:
+        "```\nREADME file is empty, please download the file and add description content.\n```",
+    tip: "tip",
+    summary: "Summary",
+    files: "Files",
+    community: "Community",
+    settings: "Settings",
+    tasks: "Tasks",
+    framework: "Framework",
+    others: "Others",
+    contributors: "Contributors",
+    historyCommits: "History commits",
+    addFile: "Add file",
+    fileName: "File name",
+    filterTags: "Filter tags",
+    provideMoreDesc: "Provide more description",
+    submitNewFile: "Submit new file",
+    submitChange: "Submit changes",
+    createNewFile: "Create new file",
+    createFile: "Create File",
+    uploadFile: "Upload file",
+    loadError: "Load Error",
+    notSupportPreview:
+        "Preview is not supported at the moment, please download via git clone",
+    changeVis: "Change visibility to",
+    enterPls: "Enter Please",
+    sureDel: "Sure to Del",
+    canNot: "Cannot",
+    role: "Role",
+    userName: "Username",
+    nickName: "Nickname",
+    phone: "Phone Number",
+    email: "Email",
+    confirm: "Confirm",
+    cancel: "Cancel",
+    add: "Add",
+    update: "Update",
+    changeSuccess: "Switch successful",
+    changeCancel: "Operation canceled",
+    addSuccess: "Added successfully",
+    delSuccess: "Del successfully",
+    lastTime: "Last Updated",
+    downloadCount: "Download",
+    public: "Public",
+    private: "Private",
+    recentlyUpdate: "Recently Update",
+    trending: "Trending",
+    mostDownload: "Most Download",
+    mostFavorite: "Most Favorite",
+    noData: "No Data",
+    save: "Save",
+    remove: "Remove",
+    saveSetting: "Save Setting",
+    toBeUploaded: "to be uploaded",
+    upLoadSuccess: "Upload successful",
+    upLoadError: "Upload error",
+    clickUpload: "Click Upload",
+    uploadAvatar: "Upload Avatar",
+    uploadPic1Mb: "Please upload image files not exceeding 1MB.",
+    select: "Select",
+    selectFilePls: "Please select a file",
+    fileTooLarge: "File size is too large",
+    inputPls: "Please enter",
+    inputNickNamePlc: "Enter nickname",
+    inputDescPlc: "Enter description",
+    lfsTips:
+        "Before downloading, please make sure you have installed git-lfs (https://git-lfs.com).",
+    lfsTips2:
+        "If you do not want to download the content of LFS files, please add it to the environment variables.",
+}
