+++ config/locales/en/comment.js
@@ -1,0 +1,4 @@
+export const comment = {
+    sureDel: "Confirm to delete this comment.js?",
+    createComment: "Add a comment.js",
+}
