+++ config/locales/en/community.js
@@ -1,0 +1,35 @@
+export const community = {
+    discussion: {
+        back: "Back to topic list",
+        new: "New topic",
+    },
+    discussionDetail: {
+        reply: "Reply",
+        comment: "Comment",
+        cancel: "Cancel",
+        updateSuccess: "Title update successfully!",
+        warn: "Content cannot be empty",
+        addSuccess: "Add comment.js successfully",
+    },
+    MD: {
+        edit: "Edit",
+        preview: "Preview",
+        desc: "Upload pictures, audio, video, or by dragging and dropping ",
+        click: "Click here to upload",
+        default: "Leave a comment.js",
+    },
+    newDiscussion: {
+        new: "New topic",
+        title: "Title",
+        content: "Content",
+        create: "Create a topic",
+        cancel: "Cancel",
+        warn: "Content cannot be empty",
+        success: "Create topic successfully!",
+    },
+    empty: {
+        welcome: "Welcome to discuss ",
+        desc: "This is the place to discuss and collaborate with OpenCSG!",
+        new: "New topic ",
+    },
+}
