+++ config/locales/en/datasets.js
@@ -1,0 +1,51 @@
+export const datasets = {
+    title: "Datasets",
+    placeholder: "Datasets name",
+    downloadDataset: "Download Dataset",
+    datasetName: "Dataset Name",
+    datasetNameTips:
+        "Applied to the dataset path, cannot be changed after creation",
+    namespaceDatasetName: "Namespace/Dataset Name",
+    datasetNickName: "Dataset Nickname",
+    datasetDesc: "Dataset Description",
+    newDataset: {
+        title: "Create New Dataset Repository",
+        titleDesc:
+            "The repository contains all model files and revision history.",
+        owner: "Owner",
+        datasetName: "Dataset Name",
+        datasetNickName: "Dataset Nickname",
+        datasetDesc: "Dataset Description",
+        public: "Public",
+        publicDesc:
+            "Anyone on the internet can see this repository. Only you (individual) or members of your organization can commit.",
+        private: "Private",
+        privateDesc:
+            "Only you (individual) or members of your organization can view and commit to this repository.",
+        tips: "After creating the model, you can use the web or Git to upload your files.",
+        tips2:
+            "Creating public datasets is currently not supported. Please contact the administrator for manual review if needed.",
+        createDataset: "Create Dataset",
+        createSuccess: "Dataset created successfully",
+    },
+    edit: {
+        tips: "Will be displayed on the dataset list page, optional",
+        tips2: "An introduction about the dataset, optional",
+        changeVisibility: "Change dataset visibility",
+        statusText: "The current dataset is",
+        status: "Status",
+        datasetVisibility: "Dataset visibility",
+        delDataset: "Delete dataset",
+        delTips: "This operation",
+        delTips2: "cannot be undone and will permanently delete",
+        delTips3: "the dataset and all its files.",
+        privateVis: "Only visible to the creator or organization members",
+        publicVis: "Visible to anyone",
+        confirmDel: "I understand and confirm deletion",
+        privateInfo:
+            ", only the creator or organization administrators can make submissions.",
+        publicInfo: ", anyone on the internet can view this dataset.",
+        needDatasetName: "Please provide a dataset alias first.",
+        needDatasetDesc: "Please provide a dataset description first.",
+    },
+}
