+++ config/locales/en/gitToken.js
@@ -1,0 +1,9 @@
+export const gitToken = {
+    title: "User Git Token",
+    desc: "Git Token is used to authorize you to perform Git-related operations",
+    refresh: "Refresh Git Token",
+    refreshWarning:
+        "After refreshing, the existing Git Token will become invalid. Are you sure you want to continue?",
+    confirm: "Continue",
+    cancelInfo: "Operation cancelled",
+}
