+++ config/locales/en/navbar.js
@@ -1,0 +1,12 @@
+export const navbar = {
+    models: "Models",
+    datasets: "Datasets",
+    profile: "Profile",
+    editProfile: "Account Settings",
+    newModel: "New Model",
+    newDataset: "New Dataset",
+    newOrganization: "New Organization",
+    logout: "Logout",
+    loginRegister: "Login/Registe",
+    login: "Login",
+}
