+++ config/locales/en/organization.js
@@ -1,0 +1,28 @@
+export const organization = {
+    orgSetting: "Organization Settings",
+    orgInfo: "Organization Information",
+    orgNameSpace: "Organization Namespace",
+    orgSpaceTips: "Once created, the namespace cannot be changed.",
+    orgNickName: "Organization Nickname",
+    orgAvatar: "Organization Avatar",
+    orgType: "Organization Type",
+    orgHomepage: "Organization Homepage",
+    orgTeam: "Team Members",
+    model: "Model",
+    dataset: "Dataset",
+    invite: {
+        title: "Invite",
+        inviteNew: "Invite New Member",
+        inviteDesc: "You are inviting a new member to the organization",
+        addSuccess: "Member added successfully",
+    },
+    newOrganization: {
+        title: "Create New Organization",
+        createOrg: "Create Organization",
+        createSuccess: "Organization created successfully",
+        inputSpaceNamePls: "Please enter the organization namespace",
+    },
+    edit: {
+        updateSuccess: "organization updated successfully",
+    },
+}
