+++ config/locales/en/profile.js
@@ -1,0 +1,20 @@
+export const profile = {
+    editProfile: "Edit Profile",
+    settings: "Settings",
+    interests: "Interests",
+    organization: "Organization",
+    menu: {
+        profile: "Profile",
+        accountInformation: "Account Information",
+        accessToken: "Access Token",
+        gitToken: "Git Token",
+        sshKey: "SSH Keys",
+        billing: "Billing",
+    },
+    edit: {
+        title: "Profile Edit",
+        uploadAvatar: "Upload Avatar",
+        removeAvatar: "Remove Avatar",
+        updateSuccess: "Profile updated successfully",
+    },
+}
