+++ config/locales/en/repo.js
@@ -1,0 +1,10 @@
+export const repo = {
+    edit: {
+        fileName: "File Name",
+        main: "Directly submit to the main branch",
+        submit: "Submit Update",
+        desc: "Provide more description",
+        update: "Update file",
+        cancel: "Cancel",
+    }
+}
