+++ config/locales/en/rule.js
@@ -1,0 +1,6 @@
+export const rule = {
+    nameRule:
+        "String of 2-70 charaters, numbers and _.- (_.-discontinuous)",
+    nameSpaceRule:
+        "String of 2-20 charaters, numbers and _",
+}
