+++ config/locales/en/sshKey.js
@@ -1,0 +1,11 @@
+export const sshKey = {
+    addSshKey: "Add SSH Key",
+    noKeyTips: "There are currently no SSH keys added to your account",
+    sshKeyName: "SSH Key Name",
+    sshKeyContent: "SSH Key Content",
+    nameWarning: "Please enter the SSH Key Name",
+    contentWarning: "Please enter the SSH Key Content",
+    sureDelKey:
+        "Are you sure you want to delete this key from your account? This action cannot be undone.",
+    delKeyName: "Delete Key 'key name'",
+}
