+++ config/locales/zh/all.js
@@ -1,0 +1,67 @@
+export const all = {
+    defaultText: "```\nREADME文件内容为空，请下载文件，补充描述内容。\n```",
+    tip: "提示",
+    summary: "介绍",
+    files: "文件",
+    community: "讨论",
+    settings: "设置",
+    tasks: "任务",
+    framework: "框架",
+    others: "其他",
+    contributors: "贡献者",
+    historyCommits: "历史提交",
+    addFile: "添加文件",
+    fileName: "文件名",
+    filterTags: "筛选标签",
+    provideMoreDesc: "提供更多描述",
+    submitNewFile: "提交新文件",
+    submitChange: "提交变更",
+    createNewFile: "创建新文件",
+    createFile: "创建文件",
+    uploadFile: "上传文件",
+    loadError: "加载数据报错",
+    notSupportPreview: "暂不支持预览，请通过 git clone 下载",
+    changeVis: "修改可见性为",
+    enterPls: "请输入",
+    sureDel: "以确认删除",
+    canNot: "无法",
+    role: "角色",
+    userName: "用户名",
+    nickName: "用户昵称",
+    phone: "手机号码",
+    email: "邮箱",
+    confirm: "确认",
+    cancel: "取消",
+    add: "添加",
+    update: "更新",
+    changeSuccess: "切换成功",
+    changeCancel: "操作取消",
+    addSuccess: "添加成功",
+    delSuccess: "删除成功",
+    lastTime: "更新时间",
+    downloadCount: "下载热度",
+    public: "公开",
+    private: "私有",
+    recentlyUpdate: "最近更新",
+    trending: "热门",
+    mostDownload: "最多下载",
+    mostFavorite: "最多喜欢",
+    noData: "暂无数据",
+    save: "保存",
+    remove: "移除",
+    saveSetting: "保存设置",
+    toBeUploaded: "待上传",
+    upLoadSuccess: "上传完成",
+    upLoadError: "上传错误",
+    clickUpload: "点击上传",
+    uploadAvatar: "上传头像",
+    uploadPic1Mb: "请上传不超过1MB的图片文件",
+    select: "选择",
+    selectFilePls: "请选择文件",
+    fileTooLarge: "文件过大",
+    inputPls: "请输入",
+    inputNickNamePlc: "请输入别名",
+    inputDescPlc: "请输入简介",
+    lfsTips: "在下载前请确认您安装了git-lfs  (https://git-lfs.com)",
+    lfsTips2: "如果您不想下载LFS文件的内容，请在环境变量中添加",
+}
