+++ config/locales/zh/comment.js
@@ -1,0 +1,4 @@
+export const comment = {
+    sureDel: "确认删除此评论？",
+    createComment: "添加评论",
+}
