+++ config/locales/zh/gitToken.js
@@ -1,0 +1,8 @@
+export const gitToken = {
+    title: "用户 Git Token",
+    desc: "Git Token 用于授权您执行 Git 相关操作",
+    refresh: "刷新 Git Token",
+    refreshWarning: "刷新之后现有的 Git Token 将会失效，确认继续吗?",
+    confirm: "继续刷新",
+    cancelInfo: "操作已取消",
+}
