+++ config/locales/zh/models.js
@@ -1,0 +1,46 @@
+export const models = {
+    title: "模型",
+    placeholder: "模型名字",
+    downloadModel: "下载模型数据",
+    download: "下载模型",
+    modelName: "模型名称",
+    datasetNameTips: "应用于模型路径，创建后不可更改",
+    namespaceModelName: "命名空间/模型名称",
+    modelNickName: "模型别名",
+    modelDesc: "模型简介",
+    newModel: {
+        title: "新建模型仓库",
+        titleDesc: "仓库包含所有的模型文件和修订的历史记录",
+        owner: "所有者",
+        modelName: "模型名称",
+        modelNickName: "模型别名",
+        modelDesc: "模型简介",
+        public: "公开",
+        publicDesc:
+            "任何互联网上的人都可以看到这个仓库。只有你（个人）或你所在组织的成员可以提交。",
+        private: "私有",
+        privateDesc: "只有你（个人）或你所在组织的成员可以看到并提交到这个仓库。",
+        tips: "创建模型后，你可以使用网页或 Git 上传你的文件。",
+        createModel: "创建模型",
+        createSuccess: "模型创建成功",
+    },
+    edit: {
+        tips: "将会显示在模型列表页面中，选填",
+        tips2: "一段关于模型的介绍，选填",
+        changeVisibility: "修改模型可见性",
+        statusText: "当前模型目前是",
+        status: "状态",
+        modelVisibility: "模型可见性",
+        delModel: "删除模型",
+        delTips: "此操作",
+        delTips2: "撤销，这将永久删除",
+        delTips3: "模型仓库及其所有文件。",
+        privateVis: "只有创建者或组织成员可见",
+        publicVis: "任何人都可以看到此模型",
+        confirmDel: "我已知晓,确认删除",
+        privateInfo: "，仅创建者或者组织管理员可进行提交",
+        publicInfo: "， 任何互联网上的人都可以看到此数据集。",
+        needModelName: "请先提供模型别名",
+        needModelDesc: "请先提供模型介绍",
+    },
+}
