+++ config/locales/zh/navbar.js
@@ -1,0 +1,12 @@
+export const navbar = {
+    models: "模型",
+    datasets: "数据集",
+    profile: "个人信息",
+    editProfile: "账号设置",
+    newModel: "新建模型",
+    newDataset: "新建数据集",
+    newOrganization: "新建组织",
+    logout: "退出登录",
+    loginRegister: "登录/注册",
+    login: "登录",
+}
