+++ config/locales/zh/organization.js
@@ -1,0 +1,28 @@
+export const organization = {
+    orgSetting: "组织设置",
+    orgInfo: "组织信息",
+    orgTeam: "团队成员",
+    orgNameSpace: "组织命名空间",
+    orgSpaceTips: "创建完成后，命名空间不可更改",
+    orgNickName: "组织别名",
+    orgAvatar: "组织头像",
+    orgType: "组织类型",
+    orgHomepage: "组织主页",
+    model: "模型",
+    dataset: "数据集",
+    invite: {
+        title: "邀请",
+        inviteNew: "邀请新成员",
+        inviteDesc: "你正在邀请新成员到组织",
+        addSuccess: "添加成员成功",
+    },
+    newOrganization: {
+        title: "新建组织",
+        createOrg: "创建组织",
+        createSuccess: "组织创建成功",
+        inputSpaceNamePls: "请填写组织命名空间",
+    },
+    edit: {
+        updateSuccess: "organization已更新",
+    },
+}
