+++ config/locales/zh/profile.js
@@ -1,0 +1,20 @@
+export const profile = {
+    editProfile: "编辑个人信息",
+    settings: "设置",
+    interests: "兴趣",
+    organization: "组织机构",
+    menu: {
+        profile: "个人资料",
+        accountInformation: "账户信息",
+        accessToken: "Access Token",
+        gitToken: "Git Token",
+        sshKey: "SSH Keys",
+        billing: "账单",
+    },
+    edit: {
+        title: "个人资料设置",
+        uploadAvatar: "上传头像",
+        removeAvatar: "移除头像",
+        updateSuccess: "profile已更新",
+    },
+}
