+++ config/locales/zh/repo.js
@@ -1,0 +1,10 @@
+export const repo = {
+    edit: {
+        fileName: "文件名",
+        main: "直接提交到 main 分支",
+        submit: "提交更新",
+        desc: "提供更多描述",
+        update: "更新文件",
+        cancel: "取消",
+    }
+}
