+++ config/locales/zh/rule.js
@@ -1,0 +1,4 @@
+export const rule = {
+    nameRule: "2-70个字母数字_.-的字符串，_.-不能并列出现",
+    nameSpaceRule: "2-20位字母数字以及 _ 构成的字符串",
+}
