+++ config/locales/zh/sshKey.js
@@ -1,0 +1,10 @@
+export const sshKey = {
+    addSshKey: "添加 SSH Key",
+    noKeyTips: "现在还没有添加SSH key到您的账户",
+    sshKeyName: "SSH Key 名称",
+    sshKeyContent: "SSH Key 内容",
+    nameWarning: "请您填写SSH Key 名称",
+    contentWarning: "请您填写SSH Key 内容",
+    sureDelKey: "你确定要从你的帐户中删除此密钥吗？此操作无法撤消。",
+    delKeyName: "删除 Key “key name”",
+}
